# make Rank() inlineable